Продолжена реализация облаков
Теперь нужно сделать GameObject-ы для всех типов облаков, и, соответственно, префабы для них
