+++ 2D_RPG/Assets/Prefabs/Clouds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 465be72e89bc11b469acc3fd1aabfaef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
