+++ 2D_RPG/Assets/Prefabs/Clouds/demo02_PixelSky_layer01.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 449d35c986dd20541bbacca037f9246f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
