+++ 2D_RPG/Assets/Prefabs/Clouds/demo02_PixelSky_layer02.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4877fb1835f359345ae5e794c041387a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
