+++ 2D_RPG/Assets/Prefabs/Clouds/demo02_PixelSky_layer03.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0e77172ab630e1541befdc6716a60c8c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
