+++ 2D_RPG/Assets/Scripts/CloudParent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba0aec3b1c984ff47982956e529279d5
+guid: 4908cbd5d15a6fa45b9c7948f0990a3d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
