+++ 2D_RPG/Assets/Scripts/NewBehaviourScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba0aec3b1c984ff47982956e529279d5
+guid: daca8a490ee11724193047bf51e33351
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
